C++17 (for now atleast)
Dropped boost since boost::filesystem is not needed anymore.
Using TOML for settings instead of boost's ini.

tbh i like c++14 for it's portability especially in a program like this, which i like to be able to ported for multiple devices easily. So i might try to get this to compile on earlier C++ standards.